make tests faster
