+++ .github/workflows/ci.yml
@@ -121,8 +121,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        shardIndex: [1, 2, 3, 4, 5, 6, 7]
-        shardTotal: [7]
+        shardIndex: [1, 2, 3, 4, 5, 6, 7, 9, 10]
+        shardTotal: [10]
     steps:
       - name: Set up Node
         uses: actions/setup-node@v5
